moved UIAlert extension from BaseVC to own file
