+++ iOS-Template/CodeDoesGood-iOSTemplate/CodeDoesGood-iOSTemplate/BaseViewController.swift
@@ -28,22 +28,3 @@ class BaseViewController: UIViewController {
 }
 
 
-extension UIAlertController {
-    func featureNotImplementedAlert() {
-        let alert = UIAlertController(title: "This feature is not yet implemented", message: nil, preferredStyle: .alert)
-        let okayAction = UIAlertAction(title: "OK", style: .cancel, handler: nil)
-        alert.addAction(okayAction)
-        self.present(alert, animated: true, completion: nil)
-    }
-    
-    func customUIAlert(title : String, message : String, buttonName: String?, action : @escaping ((UIAlertAction) -> Void?)) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-        alert.addAction(cancelAction)
-        
-        guard let buttonName = buttonName else { return }
-        
-        let customButtonAction = UIAlertAction(title: buttonName, style: .default, handler: action as? (UIAlertAction) -> Void)
-        alert.addAction(customButtonAction)
-    }
-}
